content_ai_agent/requirements.txt
crewai[tools,anthropic]==1.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytz>=2023.3
pytrends>=4.9.2
